Remove old time control picker, use new time control picker in Tournament page, and clean up time control utility functions
