+++ src/components/TimeControl/TimeControlPicker.styl
@@ -17,5 +17,5 @@
 
 
 .TimeControlPicker {
-    // here
+    
 }
